Add loading and util modules, divide backend functions into those two.
